added TextSelection class
