+++ projects/qgrid-core/infrastructure/text.selection.js
@@ -1,0 +1,31 @@
+import { Log } from './log';
+
+export class TextSelection {
+  static set(element) {  
+    element.classList.add('q-grid-can-select-text');
+  
+    if (document.body.createTextRange) {
+      const range = document.body.createTextRange();
+      range.moveToElementText(element);
+      range.select();
+    } else if (window.getSelection) {
+      const selection = window.getSelection();
+      const range = document.createRange();
+      range.selectNodeContents(element);
+      selection.removeAllRanges();
+      selection.addRange(range);
+    } else {
+      Log.error('text.selection', 'Could not select text in element: Unsupported browser.');
+    }
+  }
+  
+  static clear(element) {
+    if (window.getSelection) {
+      const selection = window.getSelection();
+      selection.removeAllRanges();	
+    }
+    if (element && element.classList) {
+      element.classList.remove('q-grid-can-select-text');
+    }
+  }
+}
