+++ projects/qgrid-core/services/dom.js
@@ -1,5 +1,4 @@
 import { isUndefined } from '../utility/kit';
-import { Log } from '../infrastructure/log';
 
 export function css(element, property, value) {
 	const normalizedProperty = normalize(property);
@@ -47,37 +46,4 @@ export function eventPath(event) {
 
 export function elementFromPoint(x, y) {
 	return document.elementFromPoint(x, y);
-}
-
-export function selectText(element) {
-	if (!element || !element.classList) {
-		Log.warn('dom.service', 'Could not select text in element: Unknown node');
-		return;
-	}
-
-	element.classList.add('q-grid-can-select-text');
-
-	if (document.body.createTextRange) {
-		const range = document.body.createTextRange();
-		range.moveToElementText(element);
-		range.select();
-	} else if (window.getSelection) {
-		const selection = window.getSelection();
-		const range = document.createRange();
-		range.selectNodeContents(element);
-		selection.removeAllRanges();
-		selection.addRange(range);
-	} else {
-		Log.warn('dom.service', 'Could not select text in element: Unsupported browser.');
-	}
-}
-
-export function removeTextSelection(element) {
-	if (window.getSelection) {
-		const selection = window.getSelection();
-		selection.removeAllRanges();	
-	}
-	if (element && element.classList) {
-		element.classList.remove('q-grid-can-select-text');
-	}
 }
